refactor: decouple nested components
