+++ .changeset/chatty-mails-happen.md
@@ -1,0 +1,5 @@
+---
+"@venusprotocol/evm": patch
+---
+
+decouple nested components
