+++ apps/evm/src/components/LabeledProgressBar/styles.tsx
