--- apps/evm/src/components/ProgressBar/PercentageRepresentation/index.stories.tsx
@@ -1,13 +1,0 @@
-import type { Meta } from '@storybook/react';
-
-import { withCenterStory } from 'stories/decorators';
-
-import { PercentageRepresentation } from '.';
-
-export default {
-  title: 'Components/PercentageRepresentation',
-  component: PercentageRepresentation,
-  decorators: [withCenterStory({ width: 100 })],
-} as Meta<typeof PercentageRepresentation>;
-
-export const PercentageRepresentationDefault = () => <PercentageRepresentation value={30} />;
