--- apps/evm/src/components/ProgressBar/PercentageRepresentation/index.tsx
@@ -1,30 +1,0 @@
-/** @jsxImportSource @emotion/react */
-import Typography from '@mui/material/Typography';
-
-import { formatPercentageToReadableValue } from 'utilities';
-
-import { ProgressBar, type ProgressBarProps } from '..';
-import { useStyles } from './styles';
-
-export const PercentageRepresentation: React.FC<Pick<ProgressBarProps, 'value' | 'className'>> = ({
-  value,
-  className,
-}) => {
-  const styles = useStyles();
-  return (
-    <div css={styles.root}>
-      <ProgressBar
-        className={className}
-        value={value}
-        mark={undefined}
-        step={1}
-        ariaLabel="percentage"
-        min={0}
-        max={100}
-      />
-      <Typography component="span" variant="small1" color="text.primary" css={styles.percentage}>
-        {formatPercentageToReadableValue(value)}
-      </Typography>
-    </div>
-  );
-};
