--- apps/evm/src/components/ProgressBar/PercentageRepresentation/styles.ts
@@ -1,16 +1,0 @@
-import { css } from '@emotion/react';
-import { useTheme } from '@mui/material';
-
-export const useStyles = () => {
-  const theme = useTheme();
-
-  return {
-    root: css`
-      display: flex;
-      align-items: center;
-    `,
-    percentage: css`
-      margin-left: ${theme.spacing(2)};
-    `,
-  };
-};
