+++ apps/evm/src/components/index.ts
@@ -8,8 +8,9 @@ export * from './Chip';
 export * from './Countdown';
 export * from './Delimiter';
 export * from './EllipseAddress';
-export * from './ApproveSteps/ApproveTokenSteps';
-export * from './ApproveSteps/ApproveDelegateSteps';
+export * from './ApproveTokenSteps';
+export * from './ApproveDelegateSteps';
+export * from './ApprovalSteps';
 export * from './Form';
 export * from './Icon';
 export * from './LabeledInlineContent';
@@ -21,12 +22,11 @@ export * from './ProgressBar';
 export * from './ProgressCircle';
 export * from './ProposalCard';
 export * from './Card';
-export * from './ProgressBar/AccountHealth';
-export * from './ProgressBar/PercentageRepresentation';
-export * from './ProgressBar/LabeledProgressBar';
+export * from './ProgressBar';
+export * from './AccountHealth';
+export * from './LabeledProgressBar';
 export * from './Select';
 export * from './Spinner';
-export * from './SuccessfulTransactionModal';
 export * from './Table';
 export * from './Tabs';
 export * from './TextField';
