--- apps/evm/src/hooks/__mocks__/useSuccessfulTransactionModal.ts
@@ -1,7 +1,0 @@
-const openSuccessfulTransactionModal = vi.fn();
-const closeSuccessfulTransactionModal = vi.fn();
-
-export default () => ({
-  openSuccessfulTransactionModal,
-  closeSuccessfulTransactionModal,
-});
